implementend json reading for highscores screen. Now reads from file scores.json (in the root) which contains an array of jsons, each containing a name and score object < logic

+++ src/main/java/nl/tudelft/semgroup4/util/PlayerScorePair.java
@@ -2,15 +2,15 @@ package nl.tudelft.semgroup4.util;
 
 public class PlayerScorePair {
     
-    private int score;
+    private long score;
     private String name;
     
-    public PlayerScorePair(String name, int score) {
+    public PlayerScorePair(String name, long score) {
         this.name = name;
         this.score = score;
     }
     
-    public int getScore() {
+    public long getScore() {
         return score;
     }
